refactor: split backup tests
